prePlayVersion

+++ Assets/TextSize.cs
@@ -5,12 +5,12 @@ using UnityEngine.UI;
 
 public class TextSize : MonoBehaviour {
 
-	public int coeff = 1;
+	public float coeff = 1;
 
 	float width = Screen.width;
 	// Use this for initialization
 	void Start () {
-		GetComponent<Text>().fontSize = (int)(width / 17f) * coeff;
+		GetComponent<Text>().fontSize = (int)((width / 17f) * coeff);
 	//	this.GetComponent<Text>().fontSize = 30;
 	}
